Detecting to patrolling at random times fix
The enemy would sometimes switch from detecting to patrolling state even when looking directly at the player.

Tweaked the logic for the detecting to patrolling switch coroutine and removed box collider collisions with the directional arrow.
